extract useFlash
