+++ packages/trace-viewer/src/ui/attachmentsTab.tsx
@@ -24,20 +24,7 @@ import { CodeMirrorWrapper, lineHeight } from '@web/components/codeMirrorWrapper
 import { isTextualMimeType } from '@isomorphic/mimeType';
 import { Expandable } from '@web/components/expandable';
 import { linkifyText } from '@web/renderUtils';
-import { clsx } from '@web/uiUtils';
-
-// flash is retriggered whenever the value changes
-function useFlash(flash: any | undefined) {
-  const [flashState, setFlashState] = React.useState(false);
-  React.useEffect(() => {
-    if (flash) {
-      setFlashState(true);
-      const timeout = setTimeout(() => setFlashState(false), 1000);
-      return () => clearTimeout(timeout);
-    }
-  }, [flash]);
-  return flashState;
-}
+import { clsx, useFlash } from '@web/uiUtils';
 
 type Attachment = AfterActionTraceEventAttachment & { traceUrl: string };
